+++ experimental/packages/opentelemetry-exporter-metrics-otlp-http/test/browser/CollectorMetricExporter.test.ts
@@ -22,10 +22,7 @@ import {
   Histogram,
 } from '@opentelemetry/api';
 import { ExportResultCode, hrTimeToNanoseconds } from '@opentelemetry/core';
-import {
-  AggregationTemporality,
-  ResourceMetrics,
-} from '@opentelemetry/sdk-metrics';
+import { ResourceMetrics } from '@opentelemetry/sdk-metrics';
 import * as assert from 'assert';
 import * as sinon from 'sinon';
 import { OTLPMetricExporter } from '../../src/platform/browser';
@@ -44,7 +41,10 @@ import {
   setUp,
   shutdown,
 } from '../metricsHelper';
-import { OTLPMetricExporterOptions } from '../../src';
+import {
+  AggregationTemporalityPreference,
+  OTLPMetricExporterOptions,
+} from '../../src';
 import { OTLPExporterConfigBase } from '@opentelemetry/otlp-exporter-base';
 import { IExportMetricsServiceRequest } from '@opentelemetry/otlp-transformer';
 
@@ -102,7 +102,7 @@ describe('OTLPMetricExporter - web', () => {
       beforeEach(() => {
         collectorExporter = new OTLPMetricExporter({
           url: 'http://foo.bar.com',
-          temporalityPreference: AggregationTemporality.CUMULATIVE,
+          temporalityPreference: AggregationTemporalityPreference.CUMULATIVE,
         });
       });
 
@@ -230,7 +230,7 @@ describe('OTLPMetricExporter - web', () => {
         (window.navigator as any).sendBeacon = false;
         collectorExporter = new OTLPMetricExporter({
           url: 'http://foo.bar.com',
-          temporalityPreference: AggregationTemporality.CUMULATIVE,
+          temporalityPreference: AggregationTemporalityPreference.CUMULATIVE,
         });
         // Overwrites the start time to make tests consistent
         Object.defineProperty(collectorExporter, '_startTime', {
@@ -386,7 +386,7 @@ describe('OTLPMetricExporter - web', () => {
     beforeEach(() => {
       collectorExporterConfig = {
         headers: customHeaders,
-        temporalityPreference: AggregationTemporality.CUMULATIVE,
+        temporalityPreference: AggregationTemporalityPreference.CUMULATIVE,
       };
       server = sinon.fakeServer.create();
     });
@@ -518,7 +518,7 @@ describe('when configuring via environment', () => {
     envSource.OTEL_EXPORTER_OTLP_HEADERS = 'foo=bar';
     const collectorExporter = new OTLPMetricExporter({
       headers: {},
-      temporalityPreference: AggregationTemporality.CUMULATIVE,
+      temporalityPreference: AggregationTemporalityPreference.CUMULATIVE,
     });
     assert.strictEqual(
       collectorExporter['_otlpExporter']['_headers'].foo,
@@ -531,7 +531,7 @@ describe('when configuring via environment', () => {
     envSource.OTEL_EXPORTER_OTLP_METRICS_HEADERS = 'foo=boo';
     const collectorExporter = new OTLPMetricExporter({
       headers: {},
-      temporalityPreference: AggregationTemporality.CUMULATIVE,
+      temporalityPreference: AggregationTemporalityPreference.CUMULATIVE,
     });
     assert.strictEqual(
       collectorExporter['_otlpExporter']['_headers'].foo,